+++ lib/dnsimple/registrar.ex
@@ -16,6 +16,7 @@ defmodule Dnsimple.Registrar do
   alias Dnsimple.DomainPrice
   alias Dnsimple.DomainRegistration
   alias Dnsimple.DomainRenewal
+  alias Dnsimple.DomainRestore
   alias Dnsimple.DomainTransfer
   alias Dnsimple.RegistrantChange
   alias Dnsimple.RegistrantChangeCheck
@@ -167,6 +168,55 @@ defmodule Dnsimple.Registrar do
 
     Client.get(client, url, options)
     |> Response.parse(%{"data" => %DomainRenewal{}})
+  end
+
+  @doc """
+  Restores a domain.
+
+  See:
+  - https://developer.dnsimple.com/v2/registrar/#restoreDomain
+
+  ## Examples:
+
+      client = %Dnsimple.Client{access_token: "a1b2c3d4"}
+      {:ok, response} = Dnsimple.Registrar.restore_domain(client, account_id = 1010, domain_id = "example.com")
+      {:ok, response} = Dnsimple.Registrar.restore_domain(client, account_id = 1010, domain_id = "example.com", %{premium_price: "109.00"})
+
+  """
+  @spec restore_domain(Client.t(), String.t(), String.t(), Keyword.t(), Keyword.t()) ::
+          {:ok | :error, Response.t()}
+  def restore_domain(client, account_id, domain_name, attributes \\ [], options \\ []) do
+    url = Client.versioned("/#{account_id}/registrar/domains/#{domain_name}/restores")
+
+    Client.post(client, url, attributes, options)
+    |> Response.parse(%{"data" => %DomainRestore{}})
+  end
+
+  @doc """
+  Retrieve the details of an existing domain restore.
+
+  See:
+  - https://developer.dnsimple.com/v2/registrar/#getDomainRestore
+
+  ## Examples:
+
+      client = %Dnsimple.Client{access_token: "a1b2c3d4"}
+      {:ok, response} = Dnsimple.Registrar.get_domain_restore(client, account_id = 1010, domain_id = "example.com", restore_id = 1)
+
+  """
+  @spec get_domain_restore(
+          Client.t(),
+          String.t() | integer,
+          String.t(),
+          String.t() | integer,
+          Keyword.t()
+        ) :: {:ok | :error, Response.t()}
+  def get_domain_restore(client, account_id, domain_name, restore_id, options \\ []) do
+    url =
+      Client.versioned("/#{account_id}/registrar/domains/#{domain_name}/restores/#{restore_id}")
+
+    Client.get(client, url, options)
+    |> Response.parse(%{"data" => %DomainRestore{}})
   end
 
   @doc """